telegram-bot-runner/requirements.txt
anyio==4.8.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
python-dotenv==1.0.1
python-telegram-bot==21.10
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
